requirements.txt
# FIT-FLIX RAG System Requirements
# Core Dependencies
chromadb==0.4.18
langchain==0.1.0
langchain-community==0.0.10
sentence-transformers==2.2.2
transformers==4.36.0
torch==2.1.0

# Google AI
google-generativeai==0.3.2

# Vector Database
faiss-cpu==1.7.4
numpy==1.24.3

# Text Processing
pypdf==3.17.0
python-docx==1.1.0
markdown==3.5.1

# Web Frameworks
gradio==4.8.0
streamlit==1.28.1
fastapi==0.104.1
uvicorn==0.24.0

# Environment & Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
pandas==2.1.3
requests==2.31.0
tqdm==4.66.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Logging
loguru==0.7.2

# Development
jupyter==1.0.0
ipykernel==6.26.0
black==23.11.0
flake8==6.1.0
